[AIRFLOW-387] Close SQLAlchemy sessions properly
This commit adopts the `provide_session` helper in almost the entire
codebase. This ensures session are handled and closed consistently.
In particular, this ensures we don't forget to close and thus leak
database connections.

As an additional change, the `provide_session` helper has been
extended to also rollback and close created connections under error
conditions.

As an additional helper, this commit also introduces a contextmanager
that provides the same functionality as the `provide_session`
decorator. This is helpful in cases where the scope of a session should
be smaller than the entire method where it is being used.

Closes #2739 from StephanErb/session_close